[Patchset 7] profiling: Allow to generate heap dumps.
Fix ownership of HeapTracker. Before this change, there was the following
problem: HeapTracker was owned by SocketListener (via SocketListener::Entry)
and de-allocated when the last socket for a PID went away. There are two
problems with this:

* We access GlobalCallsitesTrie from two different threads, potentially racing
  and crashing.
* We actually want to keep the data around until after we have dumped it, even
  if the last socket has already gone away. The current ownership semantics do
  not allow for that.

+++ src/profiling/memory/unwinding_unittest.cc
@@ -122,8 +122,7 @@ TEST(UnwindingTest, MAYBE_DoUnwind) {
   base::ScopedFile proc_maps(base::OpenFile("/proc/self/maps", O_RDONLY));
   base::ScopedFile proc_mem(base::OpenFile("/proc/self/mem", O_RDONLY));
   GlobalCallstackTrie callsites;
-  ProcessMetadata metadata(static_cast<uint64_t>(getpid()),
-                           std::move(proc_maps), std::move(proc_mem));
+  ProcessMetadata metadata(getpid(), std::move(proc_maps), std::move(proc_mem));
   WireMessage msg;
   auto record = GetRecord(&msg);
   AllocRecord out;
